requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-cors-headers==3.5.0
django-debug-toolbar==3.2
djangorestframework==3.12.2
factory-boy==3.1.0
Faker==5.0.1
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
Pygments==2.7.3
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
